+++ tests/test_scream.py
@@ -26,29 +26,31 @@ class TestScream_01_Scream(_TestOOC):
 
         # C0 (non-staff) screams
         self.c0.ooc('/scream Hi')
-        self.c0.assert_ooc('You screamed "Hi".', ooc_over=True)
-        self.c0.assert_ic('Hi', char_id=0, pos=self.c0.pos, showname=self.c0.showname, over=True)
-        self.c1.assert_ooc('(X) {} [{}] screamed "Hi" ({}).'
+        self.c0.assert_ooc('You screamed `Hi`.', ooc_over=True)
+        self.c0.assert_ic('Hi', char_id=0, pos=self.c0.pos, showname='[S] ' + self.c0.showname,
+                          over=True)
+        self.c1.assert_ooc('(X) {} [{}] screamed `Hi` ({}).'
                            .format(self.c0_dname, 0, 0), ooc_over=True)
         self.c1.assert_ic('Hi', over=True)
-        self.c2.assert_ooc('(X) {} [{}] screamed "Hi" ({}).'
+        self.c2.assert_ooc('(X) {} [{}] screamed `Hi` ({}).'
                            .format(self.c0_dname, 0, 0), over=True)
         self.c2.assert_no_ic()  # C2 is not in scream range
         self.c3.assert_no_packets()  # C3 is not in scream range
 
         self.c3.ooc('/scream Hu')
-        self.c0.assert_ooc('You heard {} scream nearby.'.format(self.c3_dname), ooc_over=True)
-        self.c0.assert_ic('Hu', showname=self.c3.showname,
+        self.c0.assert_ooc('You heard {} scream `Hu` nearby.'.format(self.c3_dname), ooc_over=True)
+        self.c0.assert_ic('Hu', showname='[S] ' + self.c3.showname,
                           over=True)  # C0 IS in scream range (compare to previous situation)
-        self.c1.assert_ooc('(X) {} [{}] screamed "Hu" ({}).'
+        self.c1.assert_ooc('(X) {} [{}] screamed `Hu` ({}).'
                            .format(self.c3_dname, 3, 5), ooc_over=True)
-        self.c1.assert_ic('Hu', showname=self.c3.showname,
+        self.c1.assert_ic('Hu', showname='[S] ' + self.c3.showname,
                           over=True)  # C1 IS in scream range (compare to previous situation)
-        self.c2.assert_ooc('(X) {} [{}] screamed "Hu" ({}).'
+        self.c2.assert_ooc('(X) {} [{}] screamed `Hu` ({}).'
                            .format(self.c3_dname, 3, 5), over=True)
         self.c2.assert_no_ic()  # C2 is not in scream range
-        self.c3.assert_ooc('You screamed "Hu".', ooc_over=True)
-        self.c3.assert_ic('Hu', char_id=3, pos=self.c3.pos, showname=self.c3.showname, over=True)
+        self.c3.assert_ooc('You screamed `Hu`.', ooc_over=True)
+        self.c3.assert_ic('Hu', char_id=3, pos=self.c3.pos, showname='[S] ' + self.c3.showname,
+                          over=True)
 
     def test_03_staffscream(self):
         """
@@ -57,11 +59,12 @@ class TestScream_01_Scream(_TestOOC):
 
         # C1 (staff) screams
         self.c1.ooc('/scream Ha')
-        self.c0.assert_ooc('You heard {} scream nearby.'.format(self.c1_dname), ooc_over=True)
-        self.c0.assert_ic('Ha', showname=self.c1.showname, over=True)
-        self.c1.assert_ooc('You screamed "Ha".', ooc_over=True)
-        self.c1.assert_ic('Ha', char_id=1, pos=self.c1.pos, showname=self.c1.showname, over=True)
-        self.c2.assert_ooc('(X) {} [{}] screamed "Ha" ({}).'.format(self.c1_dname, 1, 0),
+        self.c0.assert_ooc('You heard {} scream `Ha` nearby.'.format(self.c1_dname), ooc_over=True)
+        self.c0.assert_ic('Ha', showname='[S] ' + self.c1.showname, over=True)
+        self.c1.assert_ooc('You screamed `Ha`.', ooc_over=True)
+        self.c1.assert_ic('Ha', char_id=1, pos=self.c1.pos, showname='[S] ' + self.c1.showname,
+                          over=True)
+        self.c2.assert_ooc('(X) {} [{}] screamed `Ha` ({}).'.format(self.c1_dname, 1, 0),
                            over=True)
         self.c2.assert_no_ic()  # C2 is not in scream range
         self.c3.assert_no_packets()  # C3 is not in scream range
@@ -69,11 +72,12 @@ class TestScream_01_Scream(_TestOOC):
         # C2 (staff) screams
         self.c2.ooc('/scream He')
         self.c0.assert_no_packets()  # C0 is not in scream range
-        self.c1.assert_ooc('(X) {} [{}] screamed "He" ({}).'.format(self.c2_dname, 2, 4),
+        self.c1.assert_ooc('(X) {} [{}] screamed `He` ({}).'.format(self.c2_dname, 2, 4),
                            over=True)
         self.c1.assert_no_ic()  # C1 is not in scream range
-        self.c2.assert_ooc('You screamed "He".', ooc_over=True)
-        self.c2.assert_ic('He', char_id=2, pos=self.c2.pos, showname=self.c2.showname, over=True)
+        self.c2.assert_ooc('You screamed `He`.', ooc_over=True)
+        self.c2.assert_ic('He', char_id=2, pos=self.c2.pos, showname='[S] ' + self.c2.showname,
+                          over=True)
         self.c3.assert_no_packets()  # C3 is not in scream range
 
     def test_04_screamwhilemutedglobal(self):
@@ -102,8 +106,8 @@ class TestScream_01_Scream(_TestOOC):
         self.c3.ooc('/scream and let it all out')
         self.c0.assert_no_packets()
         self.c1.assert_no_packets()
-        self.c2.assert_ooc('(X) {} [{}] screamed "and let it all out" ({}).'
+        self.c2.assert_ooc('(X) {} [{}] screamed `and let it all out` ({}).'
                            .format(self.c3_dname, 3, 5), over=True)
-        self.c3.assert_ooc('You screamed "and let it all out".', ooc_over=True)
+        self.c3.assert_ooc('You screamed `and let it all out`.', ooc_over=True)
         self.c3.assert_ic('and let it all out', char_id=3, pos=self.c3.pos,
-                          showname=self.c3.showname, over=True)
+                          showname='[S] ' + self.c3.showname, over=True)