+++ tests/test_whisper.py
@@ -171,17 +171,19 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
             for identifier in self.get_identifiers(recipient):
                 message = '{} with {} to {}'.format(sender, identifier, recipient)
                 sent_ooc = 'You whispered `{}` to {}.'.format(message, recipient.displayname)
-                recipient_ooc = '{} whispered something to you.'.format(sender.displayname)
+                recipient_ooc = '{} whispered `{}` to you.'.format(sender.displayname, message)
                 staff_ooc = ('(X) {} [{}] whispered `{}` to {} [{}] ({}).'
                              .format(sender.displayname, sender.id, message, recipient.displayname,
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                 char_id=sender.char_id, showname=sender.showname, over=True)
+                                 char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                 over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                    char_id=sender.char_id, showname=sender.showname, over=True)
+                                    char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                    over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_ooc(staff_ooc, over=True)
                 self.c3.assert_no_packets()
@@ -200,7 +202,7 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
             for identifier in self.get_identifiers(recipient):
                 message = '{} with {} to {}'.format(sender, identifier, recipient)
                 sent_ooc = 'You whispered `{}` to {}.'.format(message, recipient.displayname)
-                recipient_ooc = '{} whispered something to you.'.format(sender.displayname)
+                recipient_ooc = '{} whispered `{}` to you.'.format(sender.displayname, message)
                 staff_ooc = ('(X) {} [{}] whispered `{}` to {} [{}] ({}).'
                              .format(sender.displayname, sender.id, message, recipient.displayname,
                                      recipient.id, sender.area.id))
@@ -209,10 +211,12 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                 char_id=sender.char_id, showname=sender.showname, over=True)
+                                 char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                 over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                    char_id=sender.char_id, showname=sender.showname, over=True)
+                                    char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                    over=True)
                 self.c2.assert_ooc(staff_ooc, over=True)
                 self.c3.assert_no_packets()
                 self.c4.assert_ooc(nonstaff_ooc, over=True)
@@ -231,17 +235,19 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
             for identifier in self.get_identifiers(recipient):
                 message = '{} with {} to {}'.format(sender, identifier, recipient)
                 sent_ooc = 'You whispered `{}` to {}.'.format(message, recipient.displayname)
-                recipient_ooc = '{} whispered something to you.'.format(sender.displayname)
+                recipient_ooc = '{} whispered `{}` to you.'.format(sender.displayname, message)
                 staff_ooc = ('(X) {} [{}] whispered `{}` to {} [{}] ({}).'
                              .format(sender.displayname, sender.id, message, recipient.displayname,
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                 char_id=sender.char_id, showname=sender.showname, over=True)
+                                 char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                 over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                    char_id=sender.char_id, showname=sender.showname, over=True)
+                                    char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                    over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c3.assert_no_packets()
                 self.c4.assert_no_packets()
@@ -268,14 +274,16 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                 message = '{} with {} to {}'.format(sender, identifier, recipient)
                 sent_ooc = ('You spooked {} by whispering `{}` to them while sneaking.'
                             .format(recipient.displayname, message))
-                recipient_ooc = ('You heard a whisper and you think it was directed at you, but '
-                                 'you could not seem to tell where it came from.')
+                recipient_ooc = ('You heard someone whisper `{}` and you think it was directed at '
+                                 'you, but you could not seem to tell where it came from.'
+                                 .format(message))
                 staff_ooc = ('(X) {} [{}] whispered `{}` to {} [{}] while sneaking ({}).'
                              .format(sender.displayname, sender.id, message, recipient.displayname,
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
-                sender.assert_ic(message, folder='<NOCHAR>', pos='jud', showname='???', over=True)
+                sender.assert_ic(message, folder='<NOCHAR>', pos='jud',
+                                 showname='[W] ' + sender.showname, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic(message, folder='<NOCHAR>', pos='jud', showname='???',
                                     over=True)
@@ -304,7 +312,8 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                 client.ooc('/whisper {} {}'.format(identifier, message))
                 client.assert_ooc(ooc, ooc_over=True)
                 client.assert_ic(message, folder=client.char_folder, pos=client.pos,
-                                 showname=client.showname, char_id=client.char_id, over=True)
+                                 showname='[W] ' + client.showname, char_id=client.char_id,
+                                 over=True)
 
     def test_06_whispersneakedtosneakedparty(self):
         """
@@ -346,7 +355,7 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
             for identifier in self.get_identifiers(recipient):
                 message = '{} with {} to {}'.format(sender, identifier, recipient)
                 sent_ooc = 'You whispered `{}` to {}.'.format(message, recipient.displayname)
-                recipient_ooc = '{} whispered something to you.'.format(sender.displayname)
+                recipient_ooc = '{} whispered `{}` to you.'.format(sender.displayname, message)
                 staff_ooc = ('(X) {} [{}] whispered `{}` to {} [{}] while both were sneaking and '
                              'part of the same party ({}).'
                              .format(sender.displayname, sender.id, message, recipient.displayname,
@@ -354,10 +363,12 @@ class TestWhisper_02_WhisperNormal(_TestWhisper):
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                 char_id=sender.char_id, showname=sender.showname, over=True)
+                                 char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                 over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                    char_id=sender.char_id, showname=sender.showname, over=True)
+                                    char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                    over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_no_packets()
                 self.c3.assert_no_packets()
@@ -426,11 +437,13 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic('(Your ears are ringing)', folder=sender.char_folder,
-                                 pos=sender.pos, char_id=sender.char_id, showname=sender.showname,
+                                 pos=sender.pos, char_id=sender.char_id,
+                                 showname='[W] ' + sender.showname,
                                  allow_partial_match=True, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic('(Your ears are ringing)', folder=sender.char_folder,
-                                    pos=sender.pos, char_id=sender.char_id, showname=sender.showname,
+                                    pos=sender.pos, char_id=sender.char_id,
+                                    showname='[W] ' + sender.showname,
                                     allow_partial_match=True, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_ooc(staff_ooc, over=True)
@@ -450,7 +463,7 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
         for identifier in self.get_identifiers(recipient):
             message = '{} with {} to {}'.format(sender, identifier, recipient)
             sent_ooc = 'You whispered `{}` to {}.'.format(message, recipient.displayname)
-            recipient_ooc = '{} whispered something to you.'.format(sender.displayname)
+            recipient_ooc = '{} whispered `{}` to you.'.format(sender.displayname, message)
             staff_ooc = ('(X) {} [{}] whispered `{}` to {} [{}] ({}).'
                          .format(sender.displayname, sender.id, message, recipient.displayname,
                                  recipient.id, sender.area.id))
@@ -459,11 +472,12 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
             sender.ooc('/whisper {} {}'.format(identifier, message))
             sender.assert_ooc(sent_ooc, ooc_over=True)
             sender.assert_ic('(Your ears are ringing)', folder=sender.char_folder, pos=sender.pos,
-                             char_id=sender.char_id, showname=sender.showname,
+                             char_id=sender.char_id, showname='[W] ' + sender.showname,
                              allow_partial_match=True, over=True)
             recipient.assert_ooc(recipient_ooc, ooc_over=True)
             recipient.assert_ic(message, folder=sender.char_folder, pos=sender.pos,
-                                char_id=sender.char_id, showname=sender.showname, over=True)
+                                char_id=sender.char_id, showname='[W] ' + sender.showname,
+                                over=True)
             self.c2.assert_ooc(staff_ooc, over=True)
             self.c3.assert_no_packets()
             self.c4.assert_ooc(nonstaff_ooc, over=True)
@@ -496,7 +510,8 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                                      recipient.id, sender.area.id))
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
-                sender.assert_ic(message, folder='<NOCHAR>', pos='jud', showname='???', over=True)
+                sender.assert_ic(message, folder='<NOCHAR>', pos='jud',
+                                 showname='[W] ' + sender.showname, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic('(Your ears are ringing)', folder='<NOCHAR>', pos='jud',
                                     allow_partial_match=True, showname='???', over=True)
@@ -526,7 +541,8 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                 client.ooc('/whisper {} {}'.format(identifier, message))
                 client.assert_ooc(ooc, ooc_over=True)
                 client.assert_ic('(Your ears are ringing)', folder=client.char_folder,
-                                 pos=client.pos, showname=client.showname, char_id=client.char_id,
+                                 pos=client.pos, showname='[W] ' + client.showname,
+                                 char_id=client.char_id,
                                  allow_partial_match=True, over=True)
 
     def test_05_whispersneakedtosneakedparty(self):
@@ -578,12 +594,13 @@ class TestWhisper_03_WhisperToDeafened(_TestWhisper):
                 sender.ooc('/whisper {} {}'.format(identifier, message))
                 sender.assert_ooc(sent_ooc, ooc_over=True)
                 sender.assert_ic('(Your ears are ringing)', folder=sender.char_folder,
-                                 pos=sender.pos, char_id=sender.char_id, showname=sender.showname,
+                                 pos=sender.pos, char_id=sender.char_id,
+                                 showname='[W] ' + sender.showname,
                                  allow_partial_match=True, over=True)
                 recipient.assert_ooc(recipient_ooc, ooc_over=True)
                 recipient.assert_ic('(Your ears are ringing)', folder=sender.char_folder,
                                     pos=sender.pos, char_id=sender.char_id,
-                                    showname=sender.showname,
+                                    showname='[W] ' + sender.showname,
                                     allow_partial_match=True, over=True)
                 self.c1.assert_ooc(staff_ooc, over=True)
                 self.c2.assert_no_packets()